+++ packages/json-rpc/polkadot-sdk-compat/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Add `fixChainSpec` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5539) on the PolkadotSDK node.
 - Add `fixDescendantValues` enhancer which addresses [the following issue](https://github.com/paritytech/polkadot-sdk/issues/5589) on the PolkadotSDK node.
 
+### Fixed
+
+- Fix small bug with the unpin-hash enhancer
+
 ## 2.1.0 - 2024-09-04
 
 ### Added